(PC-1546) Add properties isFinished and isFullyBooked on Offer model   + Move tests on models in sub directory   + Rename test modules of models   + Fix class method on offer   + Include new properties in routes   + Fix tests after rebase on master
