+++ models/stock.py
@@ -2,22 +2,22 @@
 from datetime import datetime, timedelta
 from pprint import pformat
 from sqlalchemy import BigInteger, \
-                       CheckConstraint, \
-                       Column, \
-                       DateTime, \
-                       DDL, \
-                       event, \
-                       ForeignKey, \
-                       Integer, \
-                       Numeric
+    CheckConstraint, \
+    Column, \
+    DateTime, \
+    DDL, \
+    event, \
+    ForeignKey, \
+    Integer, \
+    Numeric
 from sqlalchemy.orm import relationship
 
-from models.versioned_mixin import VersionedMixin
 from models.db import Model
 from models.pc_object import PcObject
 from models.providable_mixin import ProvidableMixin
 from models.soft_deletable_mixin import SoftDeletableMixin
 from utils.logger import logger
+from models.versioned_mixin import VersionedMixin
 
 
 class Stock(PcObject,
@@ -81,10 +81,16 @@ class Stock(PcObject,
         return api_errors
 
     @property
+    def isBookable(self):
+        return self.bookingLimitDatetime is None \
+               or self.bookingLimitDatetime >= datetime.utcnow()
+
+    @property
     def resolvedOffer(self):
         return self.offer or self.eventOccurrence.offer
 
-    def queryNotSoftDeleted():
+    @classmethod
+    def queryNotSoftDeleted(cls):
         return Stock.query.filter_by(isSoftDeleted=False)
 
     @staticmethod
