+++ tests/domain_offers_test.py
@@ -12,54 +12,49 @@ offer.thingId = 20
 
 
 @pytest.mark.standalone
-def test_check_digital_offer_consistency_raises_an_error_for_physical_venue_and_digital_thing():
-    # given
-    venue = Venue(from_dict={'isVirtual': False})
-    find_thing.return_value = Thing(from_dict={'url': 'https://zerlngzergner.fr'})
+class CheckDigitalOfferConsistencyTest:
+    def test_raises_an_error_for_physical_venue_and_digital_thing(self):
+        # given
+        venue = Venue(from_dict={'isVirtual': False})
+        find_thing.return_value = Thing(from_dict={'url': 'https://zerlngzergner.fr'})
 
-    # when
-    with pytest.raises(InconsistentOffer) as e:
-        check_digital_offer_consistency(offer, venue, find_thing=find_thing)
+        # when
+        with pytest.raises(InconsistentOffer) as e:
+            check_digital_offer_consistency(offer, venue, find_thing=find_thing)
 
-    # then
-    assert str(e.value) == 'Offer.venue is not virtual but Offer.thing has an URL'
+        # then
+        assert str(e.value) == 'Offer.venue is not virtual but Offer.thing has an URL'
 
+    def test_does_not_raise_an_error_for_virtual_venue_and_digital_thing(self):
+        # given
+        venue = Venue(from_dict={'isVirtual': True})
+        find_thing.return_value = Thing(from_dict={'url': 'https://zerlngzergner.fr'})
 
-@pytest.mark.standalone
-def test_check_digital_offer_consistency_does_not_raise_an_error_for_virtual_venue_and_digital_thing():
-    # given
-    venue = Venue(from_dict={'isVirtual': True})
-    find_thing.return_value = Thing(from_dict={'url': 'https://zerlngzergner.fr'})
+        # when
+        result = check_digital_offer_consistency(offer, venue, find_thing=find_thing)
 
-    # when
-    result = check_digital_offer_consistency(offer, venue, find_thing=find_thing)
+        # then
+        assert result is None
 
-    # then
-    assert result is None
+    def test_raises_an_error_for_virtual_venue_and_physical_thing(self):
+        # given
+        venue = Venue(from_dict={'isVirtual': True})
+        find_thing.return_value = Thing(from_dict={'url': None})
 
+        # when
+        with pytest.raises(InconsistentOffer) as e:
+            check_digital_offer_consistency(offer, venue, find_thing=find_thing)
 
-@pytest.mark.standalone
-def test_check_digital_offer_consistency_raises_an_error_for_virtual_venue_and_physical_thing():
-    # given
-    venue = Venue(from_dict={'isVirtual': True})
-    find_thing.return_value = Thing(from_dict={'url': None})
+        # then
+        assert str(e.value) == 'Offer.venue is virtual but Offer.thing does not have an URL'
 
-    # when
-    with pytest.raises(InconsistentOffer) as e:
-        check_digital_offer_consistency(offer, venue, find_thing=find_thing)
+    def test_does_not_raise_an_error_for_physical_venue_and_physical_thing(self):
+        # given
+        venue = Venue(from_dict={'isVirtual': False})
+        find_thing.return_value = Thing(from_dict={'url': None})
 
-    # then
-    assert str(e.value) == 'Offer.venue is virtual but Offer.thing does not have an URL'
+        # when
+        result = check_digital_offer_consistency(offer, venue, find_thing=find_thing)
 
-
-@pytest.mark.standalone
-def test_check_digital_offer_consistency_does_not_raise_an_error_for_physical_venue_and_physical_thing():
-    # given
-    venue = Venue(from_dict={'isVirtual': False})
-    find_thing.return_value = Thing(from_dict={'url': None})
-
-    # when
-    result = check_digital_offer_consistency(offer, venue, find_thing=find_thing)
-
-    # then
-    assert result is None
+        # then
+        assert result is None
