+++ tests/models/booking_test.py
