+++ tests/models/offer_type_test.py
