+++ tests/models/stock_test.py
