+++ tests/routes/get_offer_test.py
@@ -1,12 +1,11 @@
 import pytest
-from pprint import pprint
 
 from models import PcObject
 from tests.conftest import clean_database, TestClient
 from tests.test_utils import API_URL, \
     create_offerer, \
     create_user, \
-    create_venue, create_thing_offer, create_bank_information
+    create_venue, create_thing_offer, create_bank_information, create_stock
 from utils.human_ids import humanize
 
 
@@ -20,17 +19,18 @@ class Get:
             offerer = create_offerer()
             venue = create_venue(offerer)
             offer = create_thing_offer(venue)
-            venue_bank_information = create_bank_information(venue=venue, id_at_providers=venue.siret)
-            offerer_bank_information = create_bank_information(offerer=offerer, id_at_providers=offerer.siren)
+            stock = create_stock(offer=offer)
+            create_bank_information(venue=venue, id_at_providers=venue.siret)
+            create_bank_information(offerer=offerer, id_at_providers=offerer.siren)
 
-            PcObject.check_and_save(user, offer)
+            PcObject.check_and_save(user, stock)
 
             # when
-            response = TestClient().with_auth(email='user@test.com').get(API_URL + f'/offers/{humanize(offer.id)}')
+            response = TestClient().with_auth(email='user@test.com')\
+                .get(API_URL + f'/offers/{humanize(offer.id)}')
 
             # then
             response_json = response.json()
-            pprint(response_json)
             assert response.status_code == 200
             assert 'iban' in response_json['venue']
             assert 'bic' in response_json['venue']
